+++ tools/docgen/templates/markdown-templates.go
