+++ tools/docgen/templates/markdown/base-template
