+++ tools/docgen/templates/markdown/composite-full-template
