+++ tools/docgen/templates/markdown/composite-members-template
