+++ tools/docgen/templates/markdown/composite-template
