+++ tools/docgen/templates/markdown/enum-case-template
