+++ tools/docgen/templates/markdown/enum-template
