+++ tools/docgen/templates/markdown/event-template
