+++ tools/docgen/templates/markdown/function-template
