+++ tools/docgen/templates/markdown/initializer-template
