+++ tools/docgen/templates/templates.go
